% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{create_dummy_users}
\alias{create_dummy_users}
\title{Create dummy users database in tibble format}
\usage{
create_dummy_users()
}
\value{
a tibble with five columns:  \code{date_created},
\code{username}, \code{password}, \code{name}, \code{email}.
}
\description{
Create dummy users database in tibble format
}
\examples{

users_base <- create_dummy_users()
str(users_base)

}
