% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{create_email}
\alias{create_email}
\title{Create Email}
\usage{
create_email(body, footer, username = NULL, name = NULL, code = NULL, ...)
}
\arguments{
\item{body}{The body content of the email.}

\item{footer}{The footer content of the email.}

\item{username}{(Optional) The username associated with the email.}

\item{name}{(Optional) The name of the recipient.}

\item{code}{(Optional) A code or verification token to be included in the email.}

\item{...}{Additional arguments to be passed to the \link[blastula:compose_email]{blastula::compose_email} function.}
}
\value{
An email object created using the \code{blastula} package.
}
\description{
This function creates an email using the \code{blastula} package by composing the email body and footer.
}
\examples{
# Create an email
email <- create_email(
  body = "Hello {name}, your username is {username}.",
  footer = "Please contact support if you have any questions.",
  username = "john123",
  name = "John Doe"
)

}
