% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run examples}
\usage{
runExample(
  example = c("basic-signin", "basic-signin-mongodb", "basic-signin-theme",
    "basic-signup", "signup-pw-reset", "test-app")
)
}
\arguments{
\item{example}{The app to launch. Options are
"basic-signin", "basic-signin-mongodb", "basic-signin-theme",
"basic-signup", "signup-pw-reset",
"test-app"}
}
\value{
No return value, a shiny app is launched.
}
\description{
Launch an example shiny app using \code{shinyAuthX} authentication modules.
Use \code{admin}/\code{admin}, \code{user1}/\code{pass1} or \code{user2}/\code{pass2} to sign in.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  runExample("basic-signin")
  runExample("basic-signin-mongodb")
  runExample("basic-signin-theme")
  runExample("basic-signup")
  runExample("signup-pw-reset")
  runExample("test-app")
}
}
