% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAuthX-package.R
\docType{package}
\name{shinyAuthX-package}
\alias{shinyAuthX}
\alias{shinyAuthX-package}
\title{shinyAuthX: User Authentication Management for Shiny Apps}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Shiny modules that simplify user authentication mechanisms for single shiny apps, offering a full suite of sign-in, sign-up, sign-out, and password recovery features with modern UI designs that are compatible with `bslib` bootstrap themes.
}
\author{
\strong{Maintainer}: Myo Minn Oo \email{dr.myominnoo@gmail.com} (\href{https://orcid.org/0000-0003-4089-016X}{ORCID})

}
\keyword{internal}
