% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{show_ui_toggle}
\alias{show_ui_toggle}
\title{Toggle ui element and fade it away}
\usage{
show_ui_toggle(id)
}
\arguments{
\item{id}{An id character that corresponds to UI's id}
}
\description{
Toggle ui element and fade it away
}
