% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signin.R
\name{signinUI}
\alias{signinUI}
\title{\code{signin} UI module}
\usage{
signinUI(
  id,
  ...,
  .header = NULL,
  .add_forgotpw = TRUE,
  .add_btn_signup = TRUE,
  cookie_expiry = 7
)
}
\arguments{
\item{id}{An ID character that corresponds with that of the server module.}

\item{...}{additional shiny UIs.}

\item{.header}{header for the sign-in panel, defaults to \code{NULL}.}

\item{.add_forgotpw}{logical to add password recovery feature, defaults to \code{TRUE}.}

\item{.add_btn_signup}{logical to add sign-up feature, defaults to \code{FALSE}.}

\item{cookie_expiry}{number of days to request browser to retain \code{signin} cookie}
}
\value{
Shiny UI sign-in panel with text inputs for \code{username} and \code{password},
\code{btn_sigin} action button, and optional features, including \code{btn_fogortpw}
action link to password recovery panel and \code{btn_signup} action button to sign-up panel
}
\description{
Shiny UI module to be used with \link{signinServer}.
}
\examples{
library(shiny)
# devtools::load_all()


# dataframe that holds usernames, passwords and other user data
users_base <- dplyr::tibble(
  date_created = Sys.time(),
  username = c("admin", "user1", "user2"),
  password = sapply(
    c("admin", "pass1", "pass2"),
    sodium::password_store
  ),
  name = c("Admin", "User One", "User Two"),
  email = c("admin@email.com", "user1@email.com", "user2@email.com")
)


ui <- fluidPage(
  # add signin panel UI function without signup or password recovery panel
  signinUI(id = "signin", .add_forgotpw = FALSE, .add_btn_signup = FALSE),
  # setup table output to show user info after signin
  verbatimTextOutput("user_table")
)

server <- function(input, output, session) {
  # call signin module supplying data frame,
  credentials <- signinServer(
    id = "signin",
    users_db = users_base,
    sodium_hashed = TRUE
  )

  output$user_table <- renderPrint({
    # use req to only render results when credentials()$user_auth is TRUE
    req(credentials()$user_auth)
    str(credentials())
  })
}

if (interactive()) shinyApp(ui = ui, server = server)
}
